obdi-opensourcedatapipeline-main/app_airflow/requirements.txt

psycopg2>=2.9.10
kafka-python>=2.2.3
pyspark >= 3.5.5
numpy>=2.0.2
pandas>=2.2.3
vnstock>=0.2.9.2.3
requests>=2.31.0
pymysql>=1.0.2
pymongo>=4.6.0
boto3>=1.34.0
netCDF4>=1.6.4
webdavclient3>=3.14.6
sunpy>=5.0.1